+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 * add `let-values`, `let*-values` and `define-values` syntax macros
 * add `exact-integer?` function
+* add `stderr` port and `current-error-port` function
 ### Bugfix
 * fix empty vector literal
 * fix edge case in nested syntax-rules when variable in parent got expanded into identifier